requirements.txt
boto3==1.24.30
botocore==1.27.30
env==0.1.0
greenlet==1.1.2
importlib-metadata==4.12.0
jmespath==1.0.1
mysql-connector-python==8.0.29
numpy==1.21.6
pandas==1.3.5
protobuf==4.21.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sql==2022.4.0
SQLAlchemy==1.4.39
typing-extensions==4.3.0
urllib3==1.26.10
zipp==3.8.1
